Pages部署 设置 > 运行时 > 兼容性标志 > nodejs_compat
